+++ app/smartbrain_3000/web/_app/immutable/entry/start.BvEC3c_m.js
@@ -1,0 +1,1 @@
+import{n as e,s as t}from"../chunks/CmzMA_eQ.js";export{t as load_css,e as start};