+++ app/smartbrain_3000/web/_app/immutable/nodes/12.B_RvEfzy.js
@@ -1,1 +1,0 @@
-import{j as e,k as t,nt as n,o as r,r as i,rt as a}from"../chunks/CXkL0Yf_.js";import{t as o}from"../chunks/RqeBYq8A.js";import"../chunks/xihTtKlq.js";import"../chunks/rAxz0bow.js";import"../chunks/CdJTZ_yZ.js";var s=e(`<p class="muted">Loading&hellip;</p>`);function c(e,c){a(c,!1),i(()=>o(`/settings/providers`)),r(),t(e,s()),n()}export{c as component};